Move field related stuff into OpenKh.Game.Field
