+++ OpenKh.Game/Events/EventPlayer.cs
@@ -1,4 +1,5 @@
 using OpenKh.Engine;
+using OpenKh.Game.Field;
 using OpenKh.Game.Infrastructure;
 using System.Collections.Generic;
 using System.Numerics;
